refactor(server): extract background scrape loop to core/scrape_loop.py (3/8)
Third step of the server.py monolith split. Moves the six functions that
collectively own "what the daemon scraper thread does + how POST
/api/scrape triggers an immediate cycle".

Functions relocated:
- is_quiet_hours()        — 12am–5:30am CST gate
- build_cache()           — JSON cache rebuild from DB (called by /api/data too)
- count_fast_companies()  — Tier 0/1 batch size
- run_scrape_async(mode)  — sync runner used by the BG loop AND /api/scrape
- bg_scrape_loop()        — the daemon target
- should_run_bg_scraper() — opt-out gate (pytest auto-disable + env vars)

Rename note: the underscore-prefix on the originals was a module-private
hint that no longer applies once they live in their own module. New
names drop the underscore. server.py re-exports the new names under the
old _name aliases so the bodies of /api/scrape, /api/scrape/status, and
the startup block don't have to change in this PR — those edits land
with the route extractions in PRs 4–8.

Imports flowing the right direction:
  scrape_loop → core.state (writes state)
  scrape_loop → core.config (DB_PATH, FAST_INTERVAL)
  scrape_loop → core.scrape_orchestrator (run_scrape)
  scrape_loop → core.scrape_status (broker)
  server      → core.scrape_loop (re-exports)
  routes (future) → core.scrape_loop (will import directly)

No cycles. server.py: 905 → 767 LOC.

Tests: 219 passed (unchanged).
